flatbuffers

Struct TableVerifier

pub struct TableVerifier<'ver, 'opts, 'buf> { /* private fields */ }

Implementations

impl<'ver, 'opts, 'buf> TableVerifier<'ver, 'opts, 'buf>

pub fn deref( &mut self, field: VOffsetT, ) -> Result<Option<usize>, InvalidFlatbuffer>

pub fn verifier(&mut self) -> &mut Verifier<'opts, 'buf>

pub fn visit_field<T: Verifiable>( self, field_name: impl Into<Cow<'static, str>>, field: VOffsetT, required: bool, ) -> Result<Self, InvalidFlatbuffer>

pub fn visit_union<Key, UnionVerifier>( self, key_field_name: impl Into<Cow<'static, str>>, key_field_voff: VOffsetT, val_field_name: impl Into<Cow<'static, str>>, val_field_voff: VOffsetT, required: bool, verify_union: UnionVerifier, ) -> Result<Self, InvalidFlatbuffer>
where Key: Follow<'buf> + Verifiable, UnionVerifier: FnOnce(<Key as Follow<'buf>>::Inner, &mut Verifier<'_, '_>, usize) -> Result<(), InvalidFlatbuffer>,

Union verification is complicated. The schemas passes this function the metadata of the union’s key (discriminant) and value fields, and a callback. The function verifies and reads the key, then invokes the callback to perform data-dependent verification.

pub fn finish(self) -> &'ver mut Verifier<'opts, 'buf>