Struct flatbuffers::WIPOffset

pub struct WIPOffset<T>(_, _);

WIPOffset contains an UOffsetT with a special meaning: it is the location of data relative to the end of an in-progress FlatBuffer. The FlatBufferBuilder uses this to track the location of objects in an absolute way. The impl of Push converts a WIPOffset into a ForwardsUOffset.

Implementations

impl<'a, T: 'a> WIPOffset<T>

pub fn new(o: UOffsetT) -> WIPOffset<T>

Create a new WIPOffset.

pub fn as_union_value(self) -> WIPOffset<UnionWIPOffset>

Return a wrapped value that brings its meaning as a union WIPOffset into the type system.

pub fn value(self) -> UOffsetT

Get the underlying value.