flatbuffers

Trait Push

pub trait Push: Sized {
    type Output;

    // Required method
    unsafe fn push(&self, dst: &mut [u8], written_len: usize);

    // Provided methods
    fn size() -> usize { ... }
    fn alignment() -> PushAlignment { ... }
}

Trait to abstract over functionality needed to write values (either owned or referenced). Used in FlatBufferBuilder and implemented for generated types.

Required Associated Types

type Output

Required Methods

unsafe fn push(&self, dst: &mut [u8], written_len: usize)

Safety

dst is aligned to Self::alignment and has length greater than or equal to Self::size

Provided Methods

fn size() -> usize

fn alignment() -> PushAlignment

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.