Trait flatbuffers::EndianScalar

pub trait EndianScalar: Sized + PartialEq + Copy + Clone {
    fn to_little_endian(self) -> Self;
    fn from_little_endian(self) -> Self;
}

Trait for values that must be stored in little-endian byte order, but might be represented in memory as big-endian. Every type that implements EndianScalar is a valid FlatBuffers scalar value.

The Rust stdlib does not provide a trait to represent scalars, so this trait serves that purpose, too.

Note that we do not use the num-traits crate for this, because it provides “too much”. For example, num-traits provides i128 support, but that is an invalid FlatBuffers type.

Required Methods

fn to_little_endian(self) -> Self

fn from_little_endian(self) -> Self